+++ sqlite3/lib/src/api/statement.dart
@@ -5,6 +5,9 @@ import 'result_set.dart';
 
 /// A prepared statement.
 abstract class PreparedStatement {
+  /// The SQL statement backing this prepared statement.
+  String get sql;
+
   /// Returns the amount of parameters in this prepared statement.
   int get parameterCount;
 
@@ -23,6 +26,19 @@ abstract class PreparedStatement {
   /// [SqliteException] will be thrown.
   void execute([List<Object?> parameters = const <Object>[]]);
 
+  /// Executes this statement, ignoring result rows if there are any.
+  ///
+  /// Unlike [execute], which binds parameters by their index, [executeMap]
+  /// binds parameters by their name.
+  /// For instance, a SQL query `SELECT :foo, @bar;` has two named parameters
+  /// (`:foo` and `@bar`) that can occur as keys in [parameters].
+  ///
+  /// If the keys in [parameters] do not match the names of parameters in this
+  /// query, an [ArgumentError] will be thrown.
+  /// If sqlite3 reports an error while running this statement, a
+  /// [SqliteException] will be thrown.
+  void executeMap(Map<String, Object?> parameters);
+
   /// Selects all rows into a [ResultSet].
   ///
   /// If the [parameters] list does not match the amount of parameters in the
@@ -31,6 +47,12 @@ abstract class PreparedStatement {
   /// If sqlite3 reports an error while running this statement, a
   /// [SqliteException] will be thrown.
   ResultSet select([List<Object?> parameters = const <Object>[]]);
+
+  /// Selects all rows into a [ResultSet].
+  ///
+  /// Similar to [executeMap], parameters are bound by their name instead of
+  /// their index.
+  ResultSet selectMap(Map<String, Object?> parameters);
 
   /// Starts selecting rows by running this prepared statement with the given
   /// [parameters].