+++ sqlite3/lib/src/impl/database.dart
@@ -1,24 +1,29 @@
 part of 'implementation.dart';
 
 class DatabaseImpl implements Database {
+  final BindingsWithLibrary _library;
   final Bindings _bindings;
 
   final Pointer<sqlite3> _handle;
   final List<PreparedStatementImpl> _statements = [];
   final List<Pointer<Void>> _furtherAllocations = [];
 
+  late final _DatabaseUpdates _updates = _DatabaseUpdates(this);
+
   bool _isClosed = false;
 
-  DatabaseImpl(this._bindings, this._handle);
+  DatabaseImpl(this._library, this._handle)
+      : this._bindings = _library.bindings;
 
   factory DatabaseImpl.open(
-    Bindings bindings,
+    BindingsWithLibrary library,
     String filename, {
     String? vfs,
     OpenMode mode = OpenMode.readWriteCreate,
     bool uri = false,
     bool? mutex,
   }) {
+    final bindings = library.bindings;
     bindingsForStore = bindings;
 
     int flags;
@@ -45,7 +50,8 @@ class DatabaseImpl implements Database {
 
     final namePtr = allocateZeroTerminated(filename);
     final outDb = allocate<Pointer<sqlite3>>();
-    final vfsPtr = vfs == null ? nullPtr<char>() : allocateZeroTerminated(vfs);
+    final vfsPtr =
+        vfs == null ? nullPtr<sqlite3_char>() : allocateZeroTerminated(vfs);
 
     final result = bindings.sqlite3_open_v2(namePtr, outDb, flags, vfsPtr);
 
@@ -62,7 +68,7 @@ class DatabaseImpl implements Database {
     }
 
     bindings.sqlite3_extended_result_codes(dbPtr, 1);
-    return DatabaseImpl(bindings, dbPtr);
+    return DatabaseImpl(library, dbPtr);
   }
 
   @override
@@ -86,6 +92,9 @@ class DatabaseImpl implements Database {
   }
 
   @override
+  Stream<SqliteUpdate> get updates => _updates.updates;
+
+  @override
   Pointer<void> get handle => _handle;
 
   @override
@@ -100,7 +109,7 @@ class DatabaseImpl implements Database {
       _ensureOpen();
 
       final sqlPtr = allocateZeroTerminated(sql);
-      final errorOut = allocate<Pointer<char>>();
+      final errorOut = allocate<Pointer<sqlite3_char>>();
 
       final result = _bindings.sqlite3_exec(
           _handle, sqlPtr, nullPtr(), nullPtr(), errorOut);
@@ -140,11 +149,29 @@ class DatabaseImpl implements Database {
   @override
   PreparedStatement prepare(String sql,
       {bool persistent = false, bool vtab = true, bool checkNoTail = false}) {
+    return _prepareInternal(sql,
+            persistent: persistent,
+            vtab: vtab,
+            maxStatements: 1,
+            checkNoTail: checkNoTail)
+        .single;
+  }
+
+  @override
+  List<PreparedStatement> prepareMultiple(String sql,
+      {bool persistent = false, bool vtab = true}) {
+    return _prepareInternal(sql, persistent: persistent, vtab: vtab);
+  }
+
+  List<PreparedStatement> _prepareInternal(String sql,
+      {bool persistent = false,
+      bool vtab = true,
+      int? maxStatements,
+      bool checkNoTail = false}) {
     _ensureOpen();
 
     final stmtOut = allocate<Pointer<sqlite3_stmt>>();
-    final pzTail =
-        checkNoTail ? allocate<Pointer<char>>() : nullPtr<Pointer<char>>();
+    final pzTail = allocate<Pointer<sqlite3_char>>();
 
     final bytes = utf8.encode(sql);
     final sqlPtr = allocateBytes(bytes);
@@ -157,64 +184,91 @@ class DatabaseImpl implements Database {
       prepFlags |= SqlPrepareFlag.SQLITE_PREPARE_NO_VTAB;
     }
 
-    int resultCode;
-    // Use prepare_v3 if supported, fall-back to prepare_v2 otherwise
-    if (_bindings.supportsOpenV3) {
-      final function = _bindings.appropriateOpenFunction
-          .cast<NativeFunction<sqlite3_prepare_v3_native>>()
-          .asFunction<sqlite3_prepare_v3_dart>();
+    final createdStatements = <PreparedStatementImpl>[];
+    var offset = 0;
 
-      resultCode = function(
-        _handle,
-        sqlPtr.cast(),
-        bytes.length,
-        prepFlags,
-        stmtOut,
-        pzTail,
-      );
-    } else {
-      assert(
-        prepFlags == 0,
-        'Used custom preparation flags, but the loaded sqlite library does not '
-        'support prepare_v3',
-      );
+    void freeIntermediateResults() {
+      stmtOut.free();
+      sqlPtr.free();
+      pzTail.free();
 
-      final function = _bindings.appropriateOpenFunction
-          .cast<NativeFunction<sqlite3_prepare_v2_native>>()
-          .asFunction<sqlite3_prepare_v2_dart>();
-
-      resultCode = function(
-        _handle,
-        sqlPtr.cast(),
-        bytes.length,
-        stmtOut,
-        pzTail,
-      );
+      for (final stmt in createdStatements) {
+        _bindings.sqlite3_finalize(stmt.handle.cast());
+      }
     }
 
-    final stmtPtr = stmtOut.value;
-    stmtOut.free();
-    sqlPtr.free();
+    while (offset < bytes.length) {
+      int resultCode;
+      // Use prepare_v3 if supported, fall-back to prepare_v2 otherwise
+      if (_library.supportsOpenV3) {
+        final function = _library.appropriateOpenFunction
+            .cast<NativeFunction<sqlite3_prepare_v3_native>>()
+            .asFunction<sqlite3_prepare_v3_dart>();
 
-    if (resultCode != SqlError.SQLITE_OK) {
-      throwException(this, resultCode, sql);
+        resultCode = function(
+          _handle,
+          sqlPtr.elementAt(offset).cast(),
+          bytes.length,
+          prepFlags,
+          stmtOut,
+          pzTail,
+        );
+      } else {
+        assert(
+          prepFlags == 0,
+          'Used custom preparation flags, but the loaded sqlite library does '
+          'not support prepare_v3',
+        );
+
+        final function = _library.appropriateOpenFunction
+            .cast<NativeFunction<sqlite3_prepare_v2_native>>()
+            .asFunction<sqlite3_prepare_v2_dart>();
+
+        resultCode = function(
+          _handle,
+          sqlPtr.elementAt(offset).cast(),
+          bytes.length,
+          stmtOut,
+          pzTail,
+        );
+      }
+
+      if (resultCode != SqlError.SQLITE_OK) {
+        freeIntermediateResults();
+        throwException(this, resultCode, sql);
+      }
+
+      final stmtPtr = stmtOut.value;
+      final endOffset = pzTail.value.address - sqlPtr.address;
+
+      final sqlForStatement = utf8.decoder.convert(bytes, offset, endOffset);
+      final stmt = PreparedStatementImpl(sqlForStatement, stmtPtr, this);
+
+      createdStatements.add(stmt);
+
+      if (createdStatements.length == maxStatements) {
+        break;
+      }
+
+      offset = endOffset;
     }
 
     if (checkNoTail) {
       final usedBytes = pzTail.value.address - sqlPtr.address;
-      pzTail.free();
 
       if (usedBytes < bytes.length) {
-        _bindings.sqlite3_finalize(stmtPtr);
+        freeIntermediateResults();
         throw ArgumentError.value(
             sql, 'sql', 'Has trailing data after the first sql statement;');
       }
     }
 
-    final stmt = PreparedStatementImpl(sql, stmtPtr, this);
+    stmtOut.free();
+    sqlPtr.free();
+    pzTail.free();
 
-    _statements.add(stmt);
-    return stmt;
+    _statements.addAll(createdStatements);
+    return createdStatements;
   }
 
   int _eTextRep(bool deterministic, bool directOnly) {
@@ -303,6 +357,7 @@ class DatabaseImpl implements Database {
     if (_isClosed) return;
 
     _isClosed = true;
+    _updates.close();
     for (final stmt in _statements) {
       stmt.dispose();
     }