+++ sqlcipher_flutter_libs/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.1
+
+- Upgrade `SQLCipher` to version `4.5.0`
+
 ## 0.5.0
 
 - Initial pub release to match the version of `sqlite3_flutter_libs`